refactor: move LoginForm and SignUpForm to components directory and update imports
