+++ frontend/src/features/auth/pages/Login.jsx
@@ -1,4 +1,4 @@
-import LoginForm from '../forms/LoginForm';
+import LoginForm from '../components/LoginForm';
 
 export default function Login() {
   return (
