+++ frontend/src/features/auth/pages/SignUp.jsx
@@ -1,4 +1,4 @@
-import SignUpForm from '../forms/SignUpForm';
+import SignUpForm from '../components/SignUpForm';
 
 export default function SignUp() {
   return (
